+++ operators/advanced_network/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Holoscan Advanced Network library Changelog
 
+## 🚧 holoscan-networking 0.3
+
+### Non-Breaking
+
+RDMA Backend:
+-   **New RDMA backend**
+     -   Added RDMA backend supporting both client/server with RDMA-CM. SEND/RECV supported currently
+
 ## 🚧 holoscan-networking 0.2
 
 **Release Date**: TBD